LayoutTests/svg/animations/animate-number-calcMode-discrete-keyTimes-expected.txt

SVG 1.1 dynamic animation tests

PASS rect.x.baseVal.value is 100
PASS rect.x.baseVal.value is 200
PASS rect.x.baseVal.value is 300
PASS successfullyParsed is true

TEST COMPLETE
